Added a kludgey DNS cache to the json-status script.
This will extract the hostname from the REMOTE_URL and look it up separately.
If the lookup succeeds, it will store it in a cache.  Cache TTL is 10m by default (adjustable).
This resolved IP is passed in the --resolve arg for curl.

This process tries to be as "foolproof" as possible, falling back to a normal curl call in most "bad" cases.
Script will also NOT use this cache if it detects a local resolver config (in resolv.conf).

(+minor adjustment to create-uuid that has it report if it's using an existing UUID file)